Replace AddressText with general text component
