+++ src/components/MobilityPlatform/ChargerStationMarkers/components/ChargerStationContent/ChargerStationContent.js
@@ -1,13 +1,10 @@
 import { Typography } from '@material-ui/core';
 import PropTypes from 'prop-types';
 import React from 'react';
-import useLocaleText from '../../../../../utils/useLocaleText';
-import AddressText from '../../../AddressText';
+import SingleValueText from '../../../SingleValueText';
 import TextComponent from '../../../TextComponent';
 
 const ChargerStationContent = ({ classes, intl, station }) => {
-  const getLocaleText = useLocaleText();
-
   const titleTypo = (messageId, props = {}) => (
     <div {...props}>
       <Typography variant="subtitle2">
@@ -46,7 +43,7 @@ const ChargerStationContent = ({ classes, intl, station }) => {
       sv: item.sv,
     };
 
-    return singleValTypo('mobilityPlatform.chargerStations.content.admin', getLocaleText(stationAdmin), { className: classes.margin });
+    return <SingleValueText messageId="mobilityPlatform.chargerStations.content.admin" textObj={stationAdmin} />;
   };
 
   const renderPayment = (paymentType, props = {}) => {
@@ -68,7 +65,7 @@ const ChargerStationContent = ({ classes, intl, station }) => {
   // key property on .map() is long but it's only way to prevent all duplicate keys -warnings.
   const chargerStationInfo = (
     <>
-      {station.address ? <AddressText addressObj={stationAddress} /> : null}
+      {station.address ? <SingleValueText messageId="mobilityPlatform.content.address" textObj={stationAddress} /> : null}
       {station.extra.administrator.fi !== '' ? renderAdministrator(station.extra.administrator) : null}
       {renderPayment(station.extra.payment, { className: classes.margin })}
       {titleTypo('mobilityPlatform.content.chargersTitle', { className: classes.margin })}
