+++ src/components/MobilityPlatform/LoadingPlaces/components/LoadingPlacesContent/LoadingPlacesContent.js
@@ -1,6 +1,6 @@
 import PropTypes from 'prop-types';
 import React from 'react';
-import AddressText from '../../../AddressText';
+import SingleValueText from '../../../SingleValueText';
 import TextComponent from '../../../TextComponent';
 
 const LoadingPlacesContent = ({ classes, item }) => {
@@ -22,7 +22,7 @@ const LoadingPlacesContent = ({ classes, item }) => {
         <TextComponent textObj={loadingPlaceName} isTitle />
       </div>
       <div className={classes.textContainer}>
-        {item.address_fi !== '' ? <AddressText addressObj={loadingPlaceAddress} /> : null}
+        {item.address_fi !== '' ? <SingleValueText messageId="mobilityPlatform.content.address" textObj={loadingPlaceAddress} /> : null}
         <TextComponent textObj={item.extra.lastauspiste} />
         <TextComponent textObj={item.extra.Saavutettavuus} />
         <TextComponent textObj={item.extra.rajoitustyyppi} />
