+++ src/components/MobilityPlatform/SingleValueText/index.js
@@ -1,6 +1,6 @@
 import { withStyles } from '@material-ui/core';
 import { injectIntl } from 'react-intl';
-import AddressText from './AddressText';
+import SingleValueText from './SingleValueText';
 import styles from './styles';
 
-export default withStyles(styles)(injectIntl(AddressText));
+export default withStyles(styles)(injectIntl(SingleValueText));
